+++ src/common/ipv4_addr_src.rs
@@ -1,20 +1,39 @@
+use dyn_clone::DynClone;
 use std::convert::TryInto;
 use std::net::Ipv4Addr;
 
 use log::debug;
 
-pub trait IPV4AddrSource {
-    /// Retrieves the IPv4 address that the source thinks we should have, or None if it can't make a determination.
+/// Represents an object with methods for figuring out the DHT node's external IPv4 address
+pub trait IPV4AddrSource: DynClone + Send {
+    /// Retrieves the IPv4 address that the source thinks we should have,
+    /// or None if it can't make a determination at this time.
+    ///
+    /// This method will be called periodically by the DHT. Implementations
+    /// should return their current best guess for the external (globally routable) IPv4 address
+    /// of the DHT.
     fn get_best_ipv4(&self) -> Option<Ipv4Addr>;
 
     /// Adds a "vote" from another node in the DHT in respose to our queries.
+    ///
+    /// DHT will call this method when it receive a "hint" from another DHT node
+    /// about our external IPv4 address. An IPV4AddrSource implementation can
+    /// use these "hints" or "votes", or ignore them.
+    ///
+    /// # Parameters
+    /// * `their_addr` - The IP address of the DHT node that we're learning this information from.
+    /// * `proposed_addr` - The external IP address that the other DHT node says we have.
     fn add_vote(&mut self, their_addr: Ipv4Addr, proposed_addr: Ipv4Addr);
 
-    /// This will get called at some regular interval - an opportunity to allow votes to decay over time.
+    /// This will get called by DHT at some regular interval. Implementations
+    /// can use it to allow old information to "decay" over time.
     fn decay(&mut self);
 }
 
+dyn_clone::clone_trait_object!(IPV4AddrSource);
+
 /// An IPV4AddrSource that always returns the same thing
+#[derive(Clone)]
 pub struct StaticIPV4AddrSource {
     ip: Ipv4Addr,
 }
@@ -33,12 +52,14 @@ impl StaticIPV4AddrSource {
     }
 }
 
+#[derive(Clone)]
 struct IPV4Vote {
     ip: Ipv4Addr,
     votes: i32,
 }
 
 /// An IPV4Source that takes a certain number of "votes" from other nodes on the network to make its decision.
+#[derive(Clone)]
 pub struct IPV4Consensus {
     min_votes: usize,
     max_votes: usize,