+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # rustydht-lib changelog
 
-## [Unreleased]
+## [v3.0.0] - 2022-01-12
 * Add a `DHTBuilder` for more easily constructing DHT objects. Add a `DHTSettingsBuilder` for more easily constructing DHTSettings objects. Revise `DHT::new()` (this is a breaking change to the public API).
 * Remove `count_buckets` method from `NodeStorage` trait. This was an implementation detail of bucket-based storage leaking into the trait, which should be more generic. This is a breaking change to the public API.
 * Add documentation to `NodeStorage`, `NodeBucketStorage`, and `NodeWrapper`.
@@ -23,7 +23,7 @@
 * Add MessageBuilder, a fluent interface for building Message structs. Remove the old create_ methods for creating Messages. This change makes breaking changes to the public API, and is the reason for the major version bump.
 * Add an example called `dht_node` to the examples/ folder. It runs a DHT node and provides a simple HTTP status page.
 
-[Unreleased]: https://github.com/raptorswing/rustydht-lib/compare/v2.1.0...main
+[v3.0.0]: https://github.com/raptorswing/rustydht-lib/compare/v2.1.0...v3.0.0
 [v2.1.0]: https://github.com/raptorswing/rustydht-lib/compare/v2.0.1...v2.1.0
 [v2.0.1]: https://github.com/raptorswing/rustydht-lib/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/raptorswing/rustydht-lib/compare/v1.0.0...v2.0.0